Dev

See merge request gtp/wholesale-web-service/backend!195

+++ app/Listeners/Dashboard/AssignPriceGroupBreakpoints.php
@@ -2,17 +2,19 @@
 
 namespace App\Listeners\Dashboard;
 
-use App\Events\Dashboard\PriceGroupBreakpointsAssigned;
 use App\PriceGroup;
 use App\PriceGroupBreakpoint;
+use App\Repositories\Contracts\CustomerPricingPolicyRepository;
 use App\Repositories\Contracts\PriceGroupBreakpointRepository;
 use App\Repositories\Contracts\PriceGroupRepository;
+use App\Repositories\Eloquent\CustomerPricingPolicyRepositoryEloquent;
+use Carbon\Carbon;
 use Crmplease\MaterialAdmin\Events\Interfaces\ResourceEventInterface;
 use Crmplease\MaterialAdmin\Events\Traits\ValidatesAction;
 use Crmplease\MaterialAdmin\Events\Traits\ValidatesNamespace;
 use Crmplease\MaterialAdmin\Events\Traits\ValidatesResource;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\DB;
 
 /**
  * AssignPriceGroupBreakpoints listener.
@@ -34,17 +36,25 @@ class AssignPriceGroupBreakpoints
     protected $priceGroupBreakpoints;
 
     /**
+     * @var CustomerPricingPolicyRepositoryEloquent
+     */
+    protected $customerPricingPolicies;
+
+    /**
      * AssignPriceGroupBreakpoints constructor.
      * @param PriceGroupRepository $priceGroupRepository
      * @param PriceGroupBreakpointRepository $priceGroupBreakpointRepository
+     * @param CustomerPricingPolicyRepository $customerPricingPolicyRepository
      */
     public function __construct(
         PriceGroupRepository $priceGroupRepository,
-        PriceGroupBreakpointRepository $priceGroupBreakpointRepository
+        PriceGroupBreakpointRepository $priceGroupBreakpointRepository,
+        CustomerPricingPolicyRepository $customerPricingPolicyRepository
     )
     {
         $this->priceGroups = $priceGroupRepository;
         $this->priceGroupBreakpoints = $priceGroupBreakpointRepository;
+        $this->customerPricingPolicies = $customerPricingPolicyRepository;
     }
 
     /**
@@ -74,35 +84,31 @@ class AssignPriceGroupBreakpoints
                 /** @var \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder $query */
                 return $query->withTrashed();
             }
-        )->find($attributes['id']);
+        )
+            ->with([
+                'priceGroupBreakpoints',
+                'customers'
+            ])
+            ->find($attributes['id']);
 
         if ($priceGroup->manual) {
             $this->priceGroupBreakpoints->destroyWhere(['price_group_id' => $priceGroup->getKey()]);
         } else {
             $items = Arr::get($params, 'priceGroupBreakpoints', []);
-
-            $priceGroupBreakpoints = new Collection();
+            $policies = [];
 
             foreach ($items as $idx => $item) {
-
                 $id = numerize($item['id'] ?? false);
                 $removing = booleanize($item['_remove'] ?? false);
 
-                if ($removing) {
-
-                    if ($id) {
-                        $this->priceGroupBreakpoints->destroy($id);
-                    }
-
+                if ($removing && $id) {
+                    $this->priceGroupBreakpoints->destroy($id);
                     continue;
                 }
 
                 $breakpoint = (integer)Arr::get($item, 'breakpoint', 0);
                 $productGroups = (array)$item['productGroups'] ?? [];
-
-                $data = [
-                    'breakpoint' => $breakpoint
-                ];
+                $data = compact('breakpoint');
 
                 if ($id) {
                     /** @var PriceGroupBreakpoint $priceGroupBreakpoint */
@@ -116,20 +122,74 @@ class AssignPriceGroupBreakpoints
                 $priceGroupBreakpoint->productGroups()->sync($productGroups);
                 $priceGroupBreakpoint->save();
 
-                $priceGroupBreakpoints->push($priceGroupBreakpoint);
+                $policies = array_merge($policies, $this->preparePolicies($productGroups, $breakpoint));
             }
 
-            event(
-                new PriceGroupBreakpointsAssigned(
-                    $priceGroup,
-                    $priceGroupBreakpoints,
-                    $attributes,
-                    $params
-                )
-            );
+            $this->updateCustomerPricingPolicies($priceGroup, $policies);
+        }
+    }
+
+    /**
+     * @param array $productGroups
+     * @param int $priceGroupBreakpoint
+     * @return array
+     */
+    protected function preparePolicies(array $productGroups, int $priceGroupBreakpoint)
+    {
+        $policies = [];
+        $productGroups = array_filter(
+            $productGroups,
+            function (array $productGroup) {
+                return $productGroup['price'] ?? false;
+            }
+        );
+
+        foreach ($productGroups as $productGroup => $productGroupValues) {
+            $policies[] = [
+                'productsRange' => $priceGroupBreakpoint,
+                'productGroup' => $productGroup,
+                'price' => $productGroupValues['price'],
+            ];
         }
 
-        return;
+        return $policies;
+    }
+
+    protected function updateCustomerPricingPolicies(PriceGroup $priceGroup, array $policies)
+    {
+        $ids = $priceGroup->customers->pluck('id');
+        $this->customerPricingPolicies->trashWhereIn('customer_id', $ids->toArray());
+
+        foreach ($ids as $customer) {
+            $this->insertCustomerPricingPolicies($policies, $customer);
+        }
+    }
+
+    protected function insertCustomerPricingPolicies(array $policies, $customer)
+    {
+        $sql = 'Insert into customer_pricing_policies (customer_id, products_range, price, product_group_id, created_at, updated_at) values';
+        $policies = $this->transformPolicies($policies, $customer);
+        $values = implode(',', $policies);
+        $insert = "{$sql} {$values};";
+
+        DB::transaction(function () use ($insert){
+            DB::unprepared($insert);
+        });
+    }
+
+    /**
+     * @param array $policies
+     * @param $customer
+     * @return array
+     */
+    protected function transformPolicies(array $policies, $customer)
+    {
+        $now = Carbon::now();
+        $policies = array_map(function ($policy) use ($customer, $now) {
+            return "({$customer}, {$policy['productsRange']}, {$policy['price']}, {$policy['productGroup']}, '{$now}', '{$now}')";
+        }, $policies);
+
+        return $policies;
     }
 
     /**
